test(db): entire suite runs against real PostgreSQL 16 — both backends green
REAPER_TEST_DATABASE_URL selects the test backend: unset -> per-tempdir
SQLite (unchanged default); set to an admin URL -> every test provisions
its own uniquely-named PostgreSQL database via db::ephemeral_test_config,
keeping parallel tests as isolated as per-file SQLite.

Dialect fixes flushed out by running 196 lib + 28 integration tests on
PostgreSQL 16.13:
- users.email_verified bound a Rust bool into an INTEGER column
  (SQLite coerces, PostgreSQL type-errors) -> bind as i64
- metric columns were REAL, which is float4 on PostgreSQL and cannot
  decode into f64 -> DOUBLE PRECISION (SQLite gives it REAL affinity,
  so behavior there is byte-identical)
- namespace/source dynamic UPDATE builders bound numeric flags as text
  params -> inline numeric literals (values are program constants)

Result: 196 lib + 28 integration green on BOTH SQLite and PostgreSQL.

+++ services/reaper-management/src/auth/users/mod.rs
@@ -41,7 +41,7 @@ impl<'a> UserRepository<'a> {
         )
         .bind(user.id.to_string())
         .bind(&user.email)
-        .bind(user.email_verified)
+        .bind(user.email_verified as i64)
         .bind(&user.password_hash)
         .bind(user.status.to_string())
         .bind(user.created_at.to_rfc3339())
